Refactor: user 정보 authStatus 분리
- Auth 컴포넌트 SignIn 컴포넌트로 변경
- 절대 경로 변경
