+++ src/types/auth.ts
@@ -1,10 +1,3 @@
-export type UserState = {
+export type AuthState = {
   authStatus: 'authorized' | 'pending' | 'unauthorized';
-  email: string;
-  name: string;
-  nickName: string;
-  temperature: number;
-  genderType: string;
-  birthYear: number;
-  profileImageUrl: string;
 };
