+++ src/types/user.ts
@@ -1,0 +1,9 @@
+export type UserState = {
+  email: string;
+  name: string;
+  nickName: string;
+  temperature: number;
+  genderType: string;
+  birthYear: number;
+  profileImageUrl: string | null;
+};
